+++ packages/data-schema/src/ai/ConversationType.ts
@@ -29,6 +29,10 @@ interface ConversationRouteGetInput {
   id: string;
 }
 
+interface ConversationRouteDeleteInput {
+  id: string;
+}
+
 interface ConversationRouteListInput {
   limit?: number;
   nextToken?: string | null;
@@ -50,17 +54,27 @@ export interface ConversationRoute {
   /**
    * @experimental
    *
+   * Deletes an existing {@link Conversation} based on ID.
+   */
+  delete: (
+    input: ConversationRouteDeleteInput,
+  ) => SingularReturnValue<Conversation>;
+  /**
+   * @experimental
+   *
    * Lists all existing {@link Conversation}s on the current conversation route.
    */
   list: (input?: ConversationRouteListInput) => ListReturnValue<Conversation>;
 }
 
 // conversation types
-interface ConversationSendMessageInput {
+interface ConversationSendMessageInputObject {
   content: ConversationSendMessageInputContent[];
   aiContext?: string | Record<string, any>;
   toolConfiguration?: ToolConfiguration;
 }
+
+export type ConversationSendMessageInput = ConversationSendMessageInputObject | string;
 
 interface ConversationListMessagesInput {
   limit?: number;
@@ -71,13 +85,18 @@ type ConversationOnMessageHandler = (message: ConversationMessage) => void;
 
 export interface Conversation {
   id: string;
+  createdAt: string;
+  updatedAt: string;
+
+  metadata?: Record<string, any>;
+  name?: string;
   /**
    * @experimental
    *
    * Sends a message to the current conversation.
    */
   sendMessage: (
-    input: ConversationSendMessageInput,
+    input: ConversationSendMessageInput | string,
   ) => SingularReturnValue<ConversationMessage>;
   /**
    * @experimental