+++ packages/integration-tests/__tests__/defined-behavior/1-patterns/add-fields.ts
@@ -5,6 +5,7 @@ import {
   buildAmplifyConfig,
   mockedGenerateClient,
   optionsAndHeaders,
+  expectGraphqlMatches,
 } from '../../utils';
 
 describe('Specify a custom field type', () => {
@@ -71,6 +72,7 @@ describe('Specify a custom field type', () => {
   });
 
   describe('Inline definition', () => {
+    // #region covers 279bf864a21cd97b
     const schema = a
       .schema({
         Post: a.model({
@@ -81,7 +83,8 @@ describe('Specify a custom field type', () => {
           content: a.string(),
         }),
       })
-      .authorization((allow) => allow.guest());
+      .authorization((allow) => allow.publicApiKey());
+    // #endregion
 
     type Schema = ClientSchema<typeof schema>;
 
@@ -149,6 +152,7 @@ describe('Specify a custom field type', () => {
   });
 
   describe('Explicit definition', () => {
+    // #region covers b7fd0a0c85d7fa05
     const schema = a
       .schema({
         Location: a.customType({
@@ -165,7 +169,8 @@ describe('Specify a custom field type', () => {
           lastKnownLocation: a.ref('Location'),
         }),
       })
-      .authorization((allow) => allow.guest());
+      .authorization((allow) => allow.publicApiKey());
+    // #endregion
 
     type Schema = ClientSchema<typeof schema>;
 
@@ -202,6 +207,7 @@ describe('Specify a custom field type', () => {
     });
 
     describe('client runtime', () => {
+      // #region covers 230bf27a1eb5947d
       Object.entries(opInputResultMap).forEach(
         ([op, { input, resultKey, result }]) => {
           test(`${op}`, async () => {
@@ -229,6 +235,7 @@ describe('Specify a custom field type', () => {
           });
         },
       );
+      // #endregion
     });
   });
 });
@@ -296,6 +303,7 @@ describe('Specify an enum field type', () => {
   // #endregion
 
   describe('Inline definition', () => {
+    // #region covers 35bf03594722226f
     const schema = a
       .schema({
         Post: a.model({
@@ -303,7 +311,8 @@ describe('Specify an enum field type', () => {
           content: a.string(),
         }),
       })
-      .authorization((allow) => allow.guest());
+      .authorization((allow) => allow.publicApiKey());
+    // #endregion
 
     type Schema = ClientSchema<typeof schema>;
 
@@ -376,10 +385,37 @@ describe('Specify an enum field type', () => {
           });
         },
       );
+
+      test('enums types are enforced on the client', async () => {
+        const { generateClient } = mockedGenerateClient([]);
+        const config = await buildAmplifyConfig(schema);
+        Amplify.configure(config);
+        const client = generateClient<Schema>();
+
+        // #region covers 53b6683a37d9bc50
+        try {
+          await client.models.Post.create({
+            content: 'hello',
+            // WORKS - value auto-completed
+            privacySetting: 'PRIVATE',
+          });
+
+          await client.models.Post.create({
+            content: 'hello',
+            // DOES NOT WORK - TYPE ERROR
+            // @ts-expect-error
+            privacySetting: 'NOT_PUBLIC',
+          });
+        } catch {
+          // runtime not under test
+        }
+        // #endregion
+      });
     });
   });
 
   describe('Explicit definition', () => {
+    // #region covers f88634036e3d5189
     const schema = a
       .schema({
         PrivacySetting: a.enum(['PRIVATE', 'FRIENDS_ONLY', 'PUBLIC']),
@@ -393,7 +429,8 @@ describe('Specify an enum field type', () => {
           privacySetting: a.ref('PrivacySetting'),
         }),
       })
-      .authorization((allow) => allow.guest());
+      .authorization((allow) => allow.publicApiKey());
+    // #endregion
 
     type Schema = ClientSchema<typeof schema>;
 
@@ -435,8 +472,10 @@ describe('Specify an enum field type', () => {
         Amplify.configure(config);
         const client = generateClient<Schema>();
 
+        // #region covers 1f73a4fff6b93118
         const enumValues = client.enums.PrivacySetting.values();
         expect(enumValues).toStrictEqual(['PRIVATE', 'FRIENDS_ONLY', 'PUBLIC']);
+        // #endregion
       });
 
       Object.entries(opInputResultMap).forEach(
@@ -465,6 +504,179 @@ describe('Specify an enum field type', () => {
             expect(data).toEqual(normalizedResult);
           });
         },
+      );
+    });
+  });
+
+  describe('Mark field as required', () => {
+    // #region covers 0d26c55b7f416673
+    const schema = a
+      .schema({
+        Todo: a.model({
+          content: a.string().required(),
+        }),
+      })
+      .authorization((allow) => allow.publicApiKey());
+    // #endregion
+
+    type Schema = ClientSchema<typeof schema>;
+
+    test('Required field is non-nullable', () => {
+      type _test = Expect<Equal<Schema['Todo']['type']['content'], string>>;
+    });
+
+    test('Required field must be provided during create', async () => {
+      const { generateClient } = mockedGenerateClient([]);
+      const config = await buildAmplifyConfig(schema);
+      Amplify.configure(config);
+      const client = generateClient<Schema>();
+
+      // @ts-expect-error
+      await expect(client.models.Todo.create({})).rejects.toThrow();
+    });
+  });
+
+  describe('Mark field as array', () => {
+    // #region covers 2b588d62751dc8ee
+    const schema = a
+      .schema({
+        Todo: a.model({
+          content: a.string().required(),
+          todos: a.string().array(),
+        }),
+      })
+      .authorization((allow) => allow.publicApiKey());
+    // #endregion
+
+    type Schema = ClientSchema<typeof schema>;
+
+    test('Array field is an array type', () => {
+      type _test = Expect<
+        Equal<
+          Schema['Todo']['type']['todos'],
+          (string | null)[] | null | undefined
+        >
+      >;
+    });
+
+    test('Accepts arrays during create', async () => {
+      const { generateClient } = mockedGenerateClient([
+        {
+          data: {
+            createTodo: {
+              id: 'some-id',
+              content: 'some-content',
+              todos: ['a', 'b', 'c'],
+            },
+          },
+        },
+      ]);
+      const config = await buildAmplifyConfig(schema);
+      Amplify.configure(config);
+      const client = generateClient<Schema>();
+
+      const { data } = await client.models.Todo.create({
+        content: 'some-content',
+        todos: ['a', 'b', 'c'],
+      });
+
+      expect(data).toEqual(
+        expect.objectContaining({
+          content: 'some-content',
+          todos: ['a', 'b', 'c'],
+        }),
+      );
+    });
+  });
+
+  describe('Mark field as array', () => {
+    // #region covers 74405b7e731fd3fd
+    const schema = a
+      .schema({
+        Todo: a.model({
+          content: a.string().default('My new Todo'),
+        }),
+      })
+      .authorization((allow) => allow.publicApiKey());
+    // #endregion
+
+    type Schema = ClientSchema<typeof schema>;
+
+    test('Array field is an array type', () => {
+      type _test = Expect<
+        Equal<Schema['Todo']['type']['content'], string | null | undefined>
+      >;
+    });
+
+    test('Allows the defaulted field to be empty during create', async () => {
+      const { generateClient } = mockedGenerateClient([
+        {
+          data: {
+            createTodo: {
+              id: 'some-id',
+              content: 'My new Todo',
+            },
+          },
+        },
+      ]);
+      const config = await buildAmplifyConfig(schema);
+      Amplify.configure(config);
+      const client = generateClient<Schema>();
+
+      // content is not required at creation time.
+      const { data } = await client.models.Todo.create({});
+
+      // non-functional assertions -- the default is provided server-side.
+      // just demonstrating that we the field is still returned normally.
+      expect(data).toEqual(
+        expect.objectContaining({
+          content: 'My new Todo',
+        }),
+      );
+    });
+
+    test('Allows a value for the field during create', async () => {
+      const { generateClient, spy } = mockedGenerateClient([
+        {
+          data: {
+            createTodo: {
+              id: 'some-id',
+              content: 'non-default value',
+            },
+          },
+        },
+      ]);
+      const config = await buildAmplifyConfig(schema);
+      Amplify.configure(config);
+      const client = generateClient<Schema>();
+
+      // content is not required at creation time.
+      const { data } = await client.models.Todo.create({
+        content: 'non-default value',
+      });
+
+      // the defaulted field exists in the request.
+      expectGraphqlMatches(
+        optionsAndHeaders(spy)[0][0].query,
+        `mutation($input: CreateTodoInput!) {
+          createTodo(input: $input) {
+            id content createdAt updatedAt
+          }
+        }`,
+      );
+
+      expect(optionsAndHeaders(spy)[0][0].variables).toEqual({
+        input: {
+          content: 'non-default value',
+        },
+      });
+
+      // non-functional assertions -- the default is provided server-side.
+      // just demonstrating that we the field is still returned normally.
+      expect(data).toEqual(
+        expect.objectContaining({
+          content: 'non-default value',
+        }),
       );
     });
   });